+++ src/controllers/schedule.js
@@ -4,6 +4,7 @@ const resp = require('../library/responses')
 
 control.getAllData = async (req, res) => {
     try {
+        if (req.data_jwt.role != 'admin') return resp(res, 401, "You are not an admin")
         let { page, limit } = req.query
         page = page ? parseInt(page) : 1
         limit = limit ? parseInt(limit) : 100
@@ -25,6 +26,7 @@ control.getAllData = async (req, res) => {
 
 control.getData = async (req, res) => {
     try {
+        if (req.data_jwt.role != 'admin') return resp(res, 401, "You are not an admin")
         const id_schedule = req.params.number
         const result = await model.getData(id_schedule)
         if (result.rowCount == 0) throw 'data not found.'
@@ -37,6 +39,7 @@ control.getData = async (req, res) => {
 
 control.addData = async (req, res) => {
     try {
+        if (req.data_jwt.role != 'admin') return resp(res, 401, "You are not an admin")
         const { id_movie, id_location, id_premier, price, date_start, date_end, times } = req.body
         const result = await model.addData({ id_movie, id_location, id_premier, price, date_start, date_end, times })
         return resp(res, 200, result)
@@ -47,6 +50,7 @@ control.addData = async (req, res) => {
 }
 control.addDataTime = async (req, res) => {
     try {
+        if (req.data_jwt.role != 'admin') return resp(res, 401, "You are not an admin")
         const { id_schedule, time } = req.body
         const result_data = await model.getData(id_schedule)
         if (result_data.rowCount == 0) throw 'data schedule not found.'
@@ -61,6 +65,7 @@ control.addDataTime = async (req, res) => {
 
 control.updateData = async (req, res) => {
     try {
+        if (req.data_jwt.role != 'admin') return resp(res, 401, "You are not an admin")
         const id_schedule = req.params.id
         const result_data = await model.getData(id_schedule)
         if (result_data.rowCount == 0) throw 'data schedule not found.'
@@ -75,6 +80,7 @@ control.updateData = async (req, res) => {
 
 control.updateDataTime = async (req, res) => {
     try {
+        if (req.data_jwt.role != 'admin') return resp(res, 401, "You are not an admin")
         const id_time_schedule = req.params.id
         const { id_schedule, time } = req.body
         const result_data = await model.getData(id_schedule)
@@ -91,6 +97,7 @@ control.updateDataTime = async (req, res) => {
 
 control.deleteData = async (req, res) => {
     try {
+        if (req.data_jwt.role != 'admin') return resp(res, 401, "You are not an admin")
         const id_schedule = req.params.id
         const result = await model.deleteAllData({ id_schedule })
         return resp(res, 200, result)
@@ -101,6 +108,7 @@ control.deleteData = async (req, res) => {
 }
 control.deleteDataTime = async (req, res) => {
     try {
+        if (req.data_jwt.role != 'admin') return resp(res, 401, "You are not an admin")
         const id_time_schedule = req.params.id
         const result = await model.deleteAllDataTime({ id_time_schedule })
         return resp(res, 200, result)
